tests/submission/PhilipTaylor/canvas/type.name.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: type.name</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.type.html">type</a>.name</h1>
<p class="desc">HTMLCanvasElement type and toString</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.canvas.type">canvas.type</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("HTMLCanvasElement type and toString");
_addTest(function(canvas, ctx) {

_assertSame(Object.prototype.toString.call(canvas), '[object HTMLCanvasElement]', "Object.prototype.toString.call(canvas)", "'[object HTMLCanvasElement]'");


});
</script>

<div id="log"></div>
</body></html>
